backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.2
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==24.1.0

# Nouveaux paquets
httpx==0.27.0
phonenumbers==8.13.17
openpyxl==3.1.5
aiosmtplib==5.1.0

# PostgreSQL (correct)
psycopg2-binary==2.9.10
psycopg[binary]==3.2.3

# Rate limiting (optionnel)
slowapi==0.1.9